+++ NEWS.md
@@ -15,6 +15,8 @@ subtitle: plm - Linear Models for Panel Data - A set of estimators and tests for
 * `vcovXX`: for first-difference models with weighting scheme `type = "HC1"`
             fix degrees of freedom used in weighting (was previously only correct 
             for `vcovBK`).
+* `predict.plm`: when argument `newdata` is a pdata.frame, the (internal) model 
+                 frame creation based on the new data is fixed.
 
 ### Others:
 * (internal): little code generalisation in vcovG and vcovBK by extracting function